requirements.txt
google-genai
pydantic
python-dotenv
streamlit
openpyxl
